Merge our changes into upstream's v24.08 tag.

The following commits which we had applied on top have been made
redundant by upstream changes.

8c51ade6b Use wireColor alpha (or rgba) as the final wireframe opacity
25f668f43 gl_PointSize calculation works for orthographic projections
ab43cd1e7 gl_PointSize calculation works when display window framing is applied
c1529ba9d (origin/bfx/22.05) Further fix for shader linkage on Linux/OpenGL
44c240f9e Workaround for shader linkage issues on Linux/Mesa

The only change I made to upstream from the above commits was in the
gl_PointSize calculation. The maths was the same apart from we were
calling abs(width*deltax) so I added an abs call to the corresponding
place in upstream. I'm not sure whether this is necessary or not.

+++ pxr/base/gf/matrix4d.h
@@ -708,9 +708,9 @@ public:
     /// explicitly convert the result to GfVec3f, if necessary.
     GfVec3f TransformDir(const GfVec3f &vec) const {
         return GfVec3f(
-            vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0],
-            vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1],
-            vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2]);
+            float( vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0] ),
+            float( vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1] ),
+            float( vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2]) );
     }
 
     /// Transforms the row vector \e vec by the matrix, returning the result.
@@ -735,9 +735,9 @@ public:
     /// explicitly convert the result to GfVec3f, if necessary.
     GfVec3f TransformAffine(const GfVec3f &vec) const {
         return GfVec3f(
-            vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0] + _mtx[3][0],
-            vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1] + _mtx[3][1],
-            vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2] + _mtx[3][2]);
+            float( vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0] + _mtx[3][0] ),
+            float( vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1] + _mtx[3][1] ),
+            float( vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2] + _mtx[3][2] ));
     }
     /// @}
